mc oracle to mongodb

+++ mybatis-plus/src/main/java/com/coderlong/mybatisplus/pojo/EmptyContainerGps.java
@@ -6,6 +6,7 @@ import com.baomidou.mybatisplus.annotation.TableId;
 import com.baomidou.mybatisplus.annotation.TableName;
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.util.Date;
 @Data
 @TableName("mc_EmptyContainer_Gps")
@@ -16,10 +17,10 @@ public class EmptyContainerGps {
     
     private String equipmentCode;
 
-    private Double longitude;
+    private BigDecimal longitude;
 
 
-    private Double latitude;
+    private BigDecimal latitude;
 
     private Long direction;
